* Move ``get_expected_str`` out of _exceptions.py, where it does not belong, to its own file in _parser/_parsing_check.py * Group all exceptions in _exceptions.py * Remove now empty file _excep.py (now redundant with _exceptions.py)
