+++ libcst/__init__.py
@@ -4,8 +4,7 @@
 # LICENSE file in the root directory of this source tree.
 
 from libcst._batched_visitor import BatchableCSTVisitor, visit_batched
-from libcst._excep import CSTLogicError
-from libcst._exceptions import MetadataException, ParserSyntaxError
+from libcst._exceptions import MetadataException, ParserSyntaxError, CSTLogicError
 from libcst._flatten_sentinel import FlattenSentinel
 from libcst._maybe_sentinel import MaybeSentinel
 from libcst._metadata_dependent import MetadataDependent
@@ -239,7 +238,6 @@ __all__ = [
     "CSTNodeT",
     "CSTTransformer",
     "CSTValidationError",
-    "CSTLogicError",
     "CSTVisitor",
     "CSTVisitorT",
     "FlattenSentinel",
