--- libcst/_excep.py
@@ -1,8 +1,0 @@
-# Copyright (c) Meta Platforms, Inc. and affiliates.
-#
-# This source code is licensed under the MIT license found in the
-# LICENSE file in the root directory of this source tree.
-
-
-class CSTLogicError(Exception):
-    pass
