+++ libcst/_nodes/expression.py
@@ -18,7 +18,7 @@ from tokenize import (
 from typing import Callable, Generator, Literal, Optional, Sequence, Union
 
 from libcst._add_slots import add_slots
-from libcst._excep import CSTLogicError
+from libcst import CSTLogicError
 from libcst._maybe_sentinel import MaybeSentinel
 from libcst._nodes.base import CSTCodegenError, CSTNode, CSTValidationError
 from libcst._nodes.internal import (
