+++ libcst/_parser/base_parser.py
@@ -28,10 +28,10 @@ from typing import Generic, Iterable, List, Sequence, TypeVar, Union
 
 from libcst._exceptions import (
     EOFSentinel,
-    get_expected_str,
     ParserSyntaxError,
     PartialParserSyntaxError,
 )
+from libcst._parser._parsing_check import get_expected_str
 from libcst._parser.parso.pgen2.generator import DFAState, Grammar, ReservedString
 from libcst._parser.parso.python.token import TokenType
 from libcst._parser.types.token import Token
