+++ libcst/_parser/conversions/statement.py
@@ -6,7 +6,7 @@
 
 from typing import Any, Dict, List, Optional, Sequence, Tuple, Type
 
-from libcst._excep import CSTLogicError
+from libcst import CSTLogicError
 from libcst._exceptions import ParserSyntaxError, PartialParserSyntaxError
 from libcst._maybe_sentinel import MaybeSentinel
 from libcst._nodes.expression import (
